fix: 折叠面板，menu组件父子通信方法修改；子组件未引入问题修复
